replace -Zmir-enable-passes=-pass by disabled_mir_passes fn
