+++ charon/src/bin/charon-driver/main.rs
@@ -31,7 +31,6 @@ mod translate;
 
 use crate::driver::{arg_values, CharonCallbacks, CharonFailure, RunCompilerNormallyCallbacks};
 use charon_lib::{logger, options};
-use itertools::Itertools;
 use std::{env, panic};
 
 fn main() {
@@ -116,52 +115,6 @@ fn main() {
     for extra_flag in options.rustc_args.iter().cloned() {
         compiler_args.push(extra_flag);
     }
-
-    let disabled_mir_passes = [
-        "AfterConstProp",
-        "AfterGVN",
-        "AfterUnreachableEnumBranching)",
-        "BeforeConstProp",
-        "CheckAlignment",
-        "CopyProp",
-        "CriticalCallEdges",
-        "DataflowConstProp",
-        "DeduplicateBlocks",
-        "DestinationPropagation",
-        "EarlyOtherwiseBranch",
-        "EnumSizeOpt",
-        "GVN",
-        "Initial",
-        "Inline",
-        "InstSimplify",
-        "JumpThreading",
-        "LowerSliceLenCalls",
-        "MatchBranchSimplification",
-        "MentionedItems",
-        "MultipleReturnTerminators",
-        "ReferencePropagation",
-        "RemoveNoopLandingPads",
-        "RemoveStorageMarkers",
-        "RemoveUnneededDrops",
-        "RemoveZsts",
-        "RenameReturnPlace",
-        "ReorderBasicBlocks",
-        "ReorderLocals",
-        "ScalarReplacementOfAggregates",
-        "SimplifyComparisonIntegral",
-        "SimplifyLocals",
-        "SingleUseConsts",
-        "UnreachableEnumBranching",
-        "UnreachablePropagation",
-    ];
-    // Disable all these mir passes.
-    compiler_args.push(format!(
-        "-Zmir-enable-passes={}",
-        disabled_mir_passes
-            .iter()
-            .map(|p| format!("-{p}"))
-            .format(",")
-    ));
 
     trace!("Compiler arguments: {:?}", compiler_args);
